refactor(story-init): re-anchor baseline-snapshot fork inside branch-initializer after ensureepicbranchref (resolves #1587)
